add swap to vbox; do not add stty -echo

+++ install/fedora/install-root.sh
@@ -2,14 +2,13 @@
 set -e
 export HOME=/root
 cp -a /etc/skel/.??* /root
-cat > /.bashrc << 'EOF'
+
+if [ -f /.dockerinit ]; then
+    cat > /.bashrc << 'EOF'
 export HOME=/root
 cd $HOME
 . /root/.bash_profile
 EOF
-
-cat > $HOME/.post.bashrc <<'EOF'
-tty --silent && stty -echo
-EOF
+fi
 
 curl –s -L https://raw.githubusercontent.com/biviosoftware/home-env/master/install.sh | bash
